requirements.txt
requests
python-dotenv
pytrends
openai>=1.37.0
